Implement true mouselook and pitch control in PsyDoom
- Added pitch handling to player movement in p_user.cpp, allowing for vertical look adjustments.
- Updated rendering logic in r_main.cpp to incorporate pitch into the view matrix calculations.
- Introduced gViewPitch and related variables to manage pitch state and interpolation.
- Enhanced player input structure in doomdef.h to support vertical look input.
- Integrated achievement sound notifications in AchievementMgr.cpp when achievements are unlocked.
- Expanded OverlayMain functionality to include interactive menu for user settings.
- Refactored OverlayMain to manage input handling and rendering of overlay preferences.
- Improved compatibility layer for texture definitions in wad_compat.h.
- Generated CMake configuration files for build system setup.

+++ game/Doom/Base/i_texcache.cpp
@@ -363,7 +363,7 @@ static texdata_t TC_CacheTexData(const texture_t& tex) {
         pHeader->height = (int16_t)h;
         
         WadCompat::getCompatLayer().generateTexturePixels(pcIdx, (uint8_t*)(pBuf + sizeof(texlump_header_t)));
-        return { pBuf, texSize };
+        return { pBuf, (size_t)texSize };
     }
 
     // Make sure the texture's lump is loaded and get the bytes
